pharmapp2/requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
cycler==0.10.0
editdistance==0.5.3
firebase==3.0.1
Flask==1.1.2
flatbuffers==1.12
fuzzywuzzy==0.18.0
gast==0.3.3
gcloud==0.18.3
gevent==21.1.2
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jwcrypto==0.8
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.16.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pylsd-nova==1.2.0
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
python-jwt==3.3.0
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
scipy==1.6.2
six==1.15.0
sseclient==0.0.27
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.3.0
